+++ src/components/grids/grid-col/grid-col.svelte
@@ -93,7 +93,7 @@
 	}
 </script>
 
-<svelte:element this={tag} class={className('myt-col', _class, propertiesCss)} {...$$restProps}>
+<svelte:element this={tag} class={className('myt-grid', _class, propertiesCss)} {...$$restProps}>
 	<!-- slot: default -->
 	<slot />
 	<!-- /slot: default -->